remove popover code from FieldTemplate and move to separate label component
